# apiLayer/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Api UMbook",
        "description": "Aqui encontra-se a documentação de toda a api de dados da UMbook, incluindo todas as rotas e toda a informação à cerca do que é precisar passar como parâmetros, bem como os resultados que se podem esperar ao executar aquela rota.\n Muitas das rotas encontram-se protegidas, não se esqueça de realizar o login na aba dos utilizadores com as suas credenciais para ter acesso às rotas que lhe são permitido aceder. Receberá um token de acesso e terá que o copiar e colar no botão 'Authorize' no canto superior direito."
    },
    "schemes": [
        "http"
    ],
    "host": "localhost:3049",
    "basePath": "/api",
...
    "paths": {
...
    "/search/{nome}":{
      "get": {
          "tags": [
            "utilizadores"
          ],
          "summary": "Pesquisar utilizadores pelo nome.",
          "description": "Dado um nome ou parte de um nome, devolve os utilizadores que contém esse nome.",
          "operationId": "searchUserByName",
          "parameters": [
            {
              "name": "nome",
              "in": "path",
              "description": "Nome a procurar.",
              "required": true,
              "type": "string"
            }
          ],
          "responses": {
            "200": {
                "description": "Lista de Utilizadores encontrados.",
                "schema":{
                  "type": "array",
                  "items":{
                      "properties":{
                        "idUtilizador":{
                          "type":"string"
                        },
                        "nome":{
                          "type":"string"
                        }
                      }
                  }
                }
            },
            "401":{
              "description": "Falta-lhe o token de acesso ou o token não é válido."
            }
          },
          "security": [
            {
              "api_key": []
            }
          ]
      }
  },
...
    },
      "securityDefinitions":{
        "api_key":{
          "type": "apiKey",
          "in": "query",
          "name": "token"
        }
      },
      "security": [{
         "api_key": []
      }
    ],
...
}